Move ViewerActions into viewer sidebar
